Refer to our fork's module as DataDog/go-tuf
We were effectively masquerading as theupdateframework/go-tuf by
reporting that as our module path in go.mod. This meant we could only
reference our fork with a replace statement, rather than explicitly
depend on our fork. This is problematic because we have a repository in
the agent that we export to dd-trace-go, which is in turn used in
customer applications. We want them to be using the fork and NOT
upstream, but our replace statement in the agent does not apply to
user's code unless they add the replace statement, which is not a good
user experience.

Not only that, there's a behavior with `go get -u` that, when run on a
repo that relies on dd-trace-go, causes all deps to be upgraded to the
latest compatible version. This was causing go-tuf to jump from v0.3
(because customers code won't see the replace directive to our fork) to
v0.5.1 as `go get -u` does not respect the v0 unstable semantics. v0.5.1
does not compile with our downstream code and this was breaking user's
applications. There is a workaround to explicitly update only the
dd-trace-go package to the new version, but it's easy to forget and land
in a situation where the tracer fails to compile, which is also a bad
user experience.

While this does make pulling in upstream fixes challenging, as their
import statements will conflict with ours, we don't intend to stay on
this fork and continually update it. We're going to eventually land
upstream fixes and get back to tracking upstream.